btf: explicitly pass base to Handle.Spec
When parsing kernel module BTF we currently obtain a copy of the
vmlinux types. This is nice from an API standpoint: the user doesn't
have to care that they are dealing with a module. The problem is that
copying vmlinux BTF is very slow due to it's size.

We need to copy vmlinux BTF since it's otherwise possible for callers
of Handle.Spec to modify the internal copy of vmlinux BTF by accident.

Add an explicit base so that we need to copy less.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>